+++ README.md
@@ -11,33 +11,7 @@ Slang is designed as an educational project to demonstrate language implementati
 - Execution via a virtual machine
 - Interactive REPL
 - Support for primitive types (integers, floats, booleans, strings)
-- Functions and structured programming
-- Struct definitions
-
-## Project Structure
-
-The project is organized into the following crates:
-
-- `frontend`: Handles lexing, parsing, AST construction, and type checking
-- `ir`: Defines the intermediate representation (AST) and visitor pattern
-- `types`: Contains type system definitions and utilities
-- `backend`: Implements the compiler and virtual machine for executing bytecode
-- `derive`: Provides procedural macros for code generation and boilerplate reduction
-
-The codebase follows a modular architecture with clear separation of concerns:
-
-```plaintext
-slang/
-├── crate/
-│   ├── backend/                # Compilation to bytecode and VM execution
-│   ├── compilation_context/    # Compilation context to store symbols and types from the currently compilation unit
-│   ├── derive/                 # Procedural macros for code generation
-│   ├── frontend/               # Lexer, parser, and type checker
-│   ├── ir/                     # AST definitions and visitors
-│   └── types/                  # Type system
-├── src/                        # CLI application
-└── tests/                      # End-to-end and integration tests
-```
+- Functions
 
 ## Usage
