+++ crate/shared/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "slang_compilation_context"
+name = "slang_shared"
 version = "0.1.0"
 edition = "2024"
 
@@ -7,7 +7,7 @@ edition = "2024"
 default = []
 
 [lib]
-name = "slang_compilation_context"
+name = "slang_shared"
 path = "src/lib.rs"
 
 [dependencies]
