+++ crate/shared/README.md
@@ -1,0 +1,3 @@
+# Shared
+
+Functionality shared by the front end backend. Shall be moved into its own crate(s) in the future.
