+++ crate/shared/src/compilation_context.rs
