+++ crate/shared/src/lib.rs
