+++ crate/shared/src/symbol_table.rs
