+++ docs/index.html
@@ -35,8 +35,8 @@
             <div class="crate-description">Procedural macros and derive implementations</div>
         </div>
         <div class="crate-card">
-            <a href="slang_compilation_context/index.html" class="crate-name">slang_compilation_context</a>
-            <div class="crate-description">Compilation context and shared utilities</div>
+            <a href="slang_compilation_context/index.html" class="crate-name">slang_shared</a>
+            <div class="crate-description">shared utilities</div>
         </div>
     </div>
