+++ docs/style.css
@@ -32,7 +32,7 @@ h1 {
 }
 
 .crate-name {
-    font-size: 1.2rem;
+    font-size: 1.0rem;
     font-weight: bold;
     color: #007acc;
     text-decoration: none;
