+++ tests/operators/binary/arithmetic/misc.rs
