+++ tests/operators/binary/arithmetic/mod.rs
@@ -1,5 +1,5 @@
 mod plus;
 mod minus;
-mod star;
-mod slash;
+mod divide;
+mod multiply;
 mod misc;
