+++ tests/operators/binary/arithmetic/multiply.rs
