+++ tests/syntax/mod.rs
@@ -1,4 +1,3 @@
 mod basic;
 mod comments;
 mod errors;
-mod source_location_spanning;
